+++ packages/vendure-plugin-invoices/test/test-xero.ts
@@ -1,16 +1,13 @@
-import {XeroAccountingExportStrategy} from '../src';
+import { XeroAccountingExportStrategy } from '../src';
 
 (async () => {
+  require('dotenv').config();
 
-    require('dotenv').config();
+  const xero = new XeroAccountingExportStrategy({
+    clientId: process.env.XERO_CLIENT_ID!,
+    clientSecret: process.env.XERO_CLIENT_SECRET!,
+  });
+  await xero.init();
 
-    const xero = new XeroAccountingExportStrategy({
-        clientId: process.env.XERO_CLIENT_ID!,
-        clientSecret: process.env.XERO_CLIENT_SECRET!
-    });
-    await xero.init();
-
-    await xero.exportInvoice(null as any, null as any);
-
-
-})()
+  await xero.exportInvoice(null as any, null as any);
+})();