+++ packages/vendure-plugin-limited-products/src/api/api-extensions.ts
@@ -1,0 +1,9 @@
+import gql from 'graphql-tag';
+// This is only used by codegen so it knows DateTime is a custom scalar
+
+export const shopSchemaExtensions = gql`
+  extend type Product {
+    maxQuantityPerOrder: Int
+    limitPurchasePerMultipleOf: Int
+  }
+`;