+++ packages/vendure-plugin-limited-products/vitest.config.js
@@ -4,7 +4,7 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    include: './test/limit-variants-per-order.spec.ts',
+    include: './test/*.spec.ts',
   },
   plugins: [
     // SWC required to support decorators used in test plugins