+++ packages/vendure-plugin-invoices/CHANGELOG.md
@@ -1,7 +1,11 @@
-# 4.0.0 (2024-07-29)
+# 4.0.0 (2024-08-15)
 
 - Introduced exporting invoices to external Accounting platforms
 - Added `isCreditInvoice` as column in the DB in an Invoice
+
+# 3.1.3 (2024-08-04)
+
+- Update compatibility range (#480)
 
 # 3.1.2 (2024-07-09)